packages/dsw-mailer/requirements.txt
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
Jinja2==3.1.2
MarkupSafe==2.1.1
minio==7.1.10
pathvalidate==2.5.0
psycopg2==2.9.3
PyYAML==6.0
sentry-sdk==1.7.2
tenacity==8.0.1
urllib3==1.26.10
